Move vocabulary to a dedicated module
