+++ src/lib.rs
@@ -5,9 +5,6 @@ pub mod primitives;
 pub mod regex;
 pub mod vocabulary;
 
-mod locator;
-mod processor;
-
 #[cfg(feature = "python-bindings")]
 mod python_bindings;
