+++ src/vocabulary/locator.rs
