+++ src/vocabulary/processor.rs
